Cade's Update MathGame.py
I kept the original structure for the 10 question (I call it casual) gamemode.  Beyond that, I created a fully functioning second game mode called "Timed" where the user has 60 seconds to correctly answer as many questions as possible.  This includes separate functions for my gamemode as well as a separate leaderboard and some new buttons for the GUI.  Apart from that, I got rid of some bugs and did a bit of polishing.  The only thing I didn't get to was division and how a question like (3/7) is impossible to answer correctly since the user can't type in an infinitely repeating decimal